Here is the basic database
I created the database with the 5 items we discussed (ID, Quantity, Purchase Date, Price, and Description)

I created in the database table one record to make sure it actually holds data as intended. I created a record with an ID of 1 for some memory sticks at $20 each so if you are testing to see if retrieval of data from the database works that record should be available to display.